Adds separate component to join SMEView and SMEVisualization Also adapts and harmonizes title in SubmodelList/Tree
