+++ aas-web-ui/src/components/SubmodelList.vue
@@ -1,36 +1,36 @@
 <template>
     <v-container fluid class="pa-0">
         <v-card color="rgba(0,0,0,0)" elevation="0">
-            <template v-if="!singleAas || isMobile">
-                <!-- Title bar -->
-                <v-card-title :style="{ padding: isMobile ? '' : '15px 16px 16px' }">
-                    <div v-if="!selectedAAS || Object.keys(selectedAAS).length === 0">Submodel List</div>
-                    <div v-else class="d-flex align-center">
-                        <v-btn
-                            v-if="isMobile"
-                            class="ml-0"
-                            variant="plain"
-                            icon="mdi-chevron-left"
-                            @click="backToAASList()" />
+            <!-- Title bar -->
+            <v-card-title :style="{ padding: isMobile ? '' : '15px 16px 16px' }">
+                <div class="d-flex align-center">
+                    <v-btn
+                        v-if="isMobile"
+                        class="ml-0"
+                        variant="plain"
+                        icon="mdi-chevron-left"
+                        @click="backToAASList()" />
+                    <span v-if="!selectedAAS || Object.keys(selectedAAS).length === 0">Submodel List </span>
+                    <template v-else-if="!isMobile && submodelList.length > 10">
+                        <v-col class="ml-2">
+                            <v-text-field
+                                variant="outlined"
+                                density="compact"
+                                hide-details
+                                label="Search for Submodel ..."
+                                clearable
+                                @update:model-value="filterSubmodelList"></v-text-field>
+                        </v-col>
+                    </template>
+                    <template v-else>
                         <v-icon icon="custom:aasIcon" color="primary" size="small" class="ml-2" />
-                        <template v-if="submodelList.length > 10">
-                            <v-col class="ml-2">
-                                <v-text-field
-                                    variant="outlined"
-                                    density="compact"
-                                    hide-details
-                                    label="Search for Submodel ..."
-                                    clearable
-                                    @update:model-value="filterSubmodelList"></v-text-field>
-                            </v-col>
-                        </template>
-                        <span v-else class="text-truncate ml-2">
+                        <span class="text-truncate ml-2">
                             {{ nameToDisplay(selectedAAS) }}
                         </span>
-                    </div>
-                </v-card-title>
-                <v-divider></v-divider>
-            </template>
+                    </template>
+                </div>
+            </v-card-title>
+            <v-divider></v-divider>
             <v-card-text class="py-2 px-2" style="overflow-y: auto; height: calc(100svh - 170px)">
                 <div v-if="listLoading">
                     <v-skeleton-loader type="list-item@6"></v-skeleton-loader>
@@ -98,7 +98,6 @@
     import { useAASHandling } from '@/composables/AAS/AASHandling';
     import { useReferableUtils } from '@/composables/AAS/ReferableUtils';
     import { useAASStore } from '@/store/AASDataStore';
-    import { useEnvStore } from '@/store/EnvironmentStore';
     import { useNavigationStore } from '@/store/NavigationStore';
 
     // Extend the ComponentPublicInstance type to include scrollToIndex
@@ -117,7 +116,6 @@
     // Stores
     const navigationStore = useNavigationStore();
     const aasStore = useAASStore();
-    const envStore = useEnvStore();
 
     // Vuetify
     const theme = useTheme();
@@ -136,7 +134,6 @@
     const submodelRegistryURL = computed(() => navigationStore.getSubmodelRegistryURL);
     const isMobile = computed(() => navigationStore.getIsMobile);
     const primaryColor = computed(() => theme.current.value.colors.primary);
-    const singleAas = computed(() => envStore.getSingleAas); // Get the single AAS state from the Store
 
     // Watchers
     watch(
@@ -217,7 +214,9 @@
     }
 
     function selectSM(submodel: any): void {
-        if (isSelected(submodel)) {
+        // No deselection on mobile
+        // On mobile every click on a submodel routes to visualization ()
+        if (isSelected(submodel) && !isMobile.value) {
             // Deselect submodel: remove the path query
             let query = { ...route.query };
             delete query.path;
